Lag feilrespons komponent
Denne håndterer 'alle' feilkoder for revurdering & søknadsbehandling
