+++ src/components/Feilresponser/feilresponser-nb.ts
@@ -1,0 +1,49 @@
+export default {
+    'generell.fant.ikke.behandling': 'Fant ikke behandling',
+    'generell.fant.ikke.aktør.id': 'Fant ikke aktør ID',
+    'generell.fant.ikke.person': 'Fant ikke person',
+    'generell.kunne.ikke.opprette.oppgave': 'Kunne ikke opprette oppgave',
+    'generell.kunne.ikke.utbetale': 'Kunne ikke utbetale',
+
+    'generell.ugyldig.body': 'Ugyldig data som ble innsendt',
+    'generell.ugyldig.tilstand': 'Ugyldig tilstand',
+
+    'periode.ugyldig.fom': 'Perioder kan kun starte første dagen i måneden',
+    'periode.ugyldig.tom': 'Perioder kan kun slutte siste dagen i morgen',
+    'periode.ugyldig.start.slutt': 'Startmåned må være før, eller lik sluttmåned',
+
+    'vurderingsperiode.overlappende.vurderingsperioder': 'Perioder kan ikke overlappe',
+    'vurderingsperiode.vurdering.utenfor.revurderingsperioden':
+        'Vurdering av vilkår må gjøres innenfor revurderingsperioden',
+
+    'uføre.uføregrad.må.være.mellom.en.og.hundre': 'Uføregraden må være mellom 1 og 100',
+    'uføre.grad.og.forventetinntekt.mangler': 'Uføregrad og/eller forventet inntekt mangler',
+    'uføre.grunnlag.og.vurdering.forskjellige':
+        'Det er ikke samsvar mellom perioden for vurdering og perioden for grunnlaget',
+
+    'bosituasjon.kunne.ikke.legge.til': 'Kunne ikke legge til bosituasjonsgrunnlaget',
+
+    'fradrag.kunne.ikke.legge.til': 'Kunne ikke legge til fradragsgrunnlag',
+    'fradrag.mangler.periode': 'Fradrag mangler periode',
+    'fradrag.utenlandsinntekt.negativt.beløp': 'Fradrag har negativt utenlandsbeløp',
+    'fradrag.utenlandsinntekt.mangler.valuta': 'Fradrag mangler valuta',
+    'fradrag.utenlandsinntekt.negativ.kurs': 'Fradrag har negativ krus',
+
+    'simulering.simulering.feilet': 'Simulering feilet',
+    'simulering.oppdrag.stengt.eller.nede':
+        'Simulering feilet. Oppdrag/UR er stengt eller nede. Prøv på nytt eller prøv igjen i Oppdrags åpningstid.',
+    'simulering.finner.ikke.person': 'Simulering feilet. Finner ikke person i TPS.',
+    'simulering.finner.ikke.kjøretidsplan':
+        'Simulering feilet. Finner ikke kjøreplansperiode for fom-dato. Kjøreplan finnes som regel bare for inneværende år.',
+    'simulering.oppdraget.finnes.ikke':
+        'Simulering feilet. Oppdraget finnes ikke, brukerens transaksjoner er sannsynligvis fjernet fra Oppdrag.',
+
+    'brev.kunne.ikke.generere': 'Kunne ikke generere brev',
+    'brev.generering.dokument.feilet': 'Feil ved generering av dokument',
+
+    'utbetaling.kunne.ikke.utbetale': 'Kunne ikke utbetale',
+    'utbetaling.kontrollsimulering.ulik.saksbehandlers.simulering':
+        'Kontrollsimuleringen er ulik saksbehandlers simulering',
+
+    'feil.ukjentFeil': 'Ukjent feil',
+};
