+++ src/components/Feilresponser/revurderingFeilresponser/RevurderingFeilresponser-nb.ts
@@ -3,27 +3,19 @@ export default {
     'generell.attestant.og.saksbehandler.kan.ikke.være.samme.person': 'Kan ikke attestere egen saksbehandling',
     'generell.feilutbetaling.støttes.ikke': 'Simulering indikerer feilutbetaling, systemstøtte for håndtering mangler.',
     'generell.fant.ikke.sak': 'Fant ikke sak',
-    'generell.fant.ikke.person': 'Fant ikke person',
-    'generell.fant.ikke.aktør.id': 'Fant ikke aktør id',
     'generell.fant.ikke.revurdering': 'Fant ikke revurdering',
     'generell.ugyldig.periode': 'Ugyldig periode',
-    'generell.ugyldig.tilstand': 'Ugyldig tilstand',
     'generell.ugyldig.årsak': 'Ugyldig årsak',
-    'generell.ugyldig.body': 'Ugyldig data som ble innsendt',
-    'generell.kunne.ikke.opprette.oppgave': 'Kunne ikke opprette oppgave',
-    'generell.kunne.ikke.utbetale': 'Kunne ikke utbetale',
     'generell.kunne.ikke.slå.opp.eps': 'Kunne ikke slå opp ektefelle eller samboer i PDL',
 
     'periode.ingenting.å.revurdere': 'Fant ingen vedtak som kan revurderes for angitt periode',
-    'periode.overlappende.vurderingsperioder': 'Perioder kan ikke overlappe',
-    'periode.vurdering.utenfor.revurderingsperioden': 'Vurdering av vilkår må gjøres innenfor revurderingsperioden',
 
     'forhåndsvarsel.allerede.forhåndsvarslet': 'Revurderingen er allerede blitt forhåndsvarslet',
     'forhåndsvarsel.kan.ikke.oppdatere.er.forhåndsvarslet': 'Kan ikke oppdatere revurdering som er forhåndsvarslet',
     'forhåndsvarsel.mangler.beslutning': 'Mangler beslutning på forhåndsvarsel',
     'forhåndsvarsel.ugyldig.valg': 'Ugyldig valg av beslutning for forhåndsvarsel',
     'forhåndsvarsel.er.besluttet': 'Revurderingen er allerede besluttet',
-    'forhåndsvarsel.ikke.varslet': 'Revurderingen er ikke varslet for å beslutte en forhåndsvarsling',
+    'forhåndsvarsel.ikke.varslet.for.å.beslutte': 'Revurderingen er ikke varslet for å beslutte en forhåndsvarsling',
     'forhåndsvarsel.feil.tilstand.for.beslutning':
         'Revurderingen er ikke i riktig tilstand for å beslutte forhåndsvarslingen',
 
@@ -49,15 +41,10 @@ export default {
     'opprettelseOgOppdatering.oppdaterer.forhåndsvarslet.revurdering':
         'Kan ikke oppdatere en revurdering som er forhåndsvarslet',
 
-    'uføre.uføregrad.må.være.mellom.en.og.hundre': 'Uføregraden må være mellom 1 og 100',
-    'uføre.grad.og.forventetinntekt.mangler': 'Uføregrad og/eller forventet inntekt mangler',
-    'uføre.grunnlag.og.vurdering.forskjellige':
-        'Det er ikke samsvar mellom perioden for vurdering og perioden for grunnlaget',
     'uføre.vurderinger.samme.resultat': 'Vurdering av vilkår må ha samme resultat for hele revurderingsperioden',
     'uføre.hele.behandlingsperioden.må.ha.vurdering': 'Hele behandlingsperioden må ha vurderinger',
     'uføre.vurderingsperiode.mangler': 'Hele eller deler av revurderingsperioden mangler vurdering(er)',
 
-    'bosituasjon.kunne.ikke.legge.til': 'Kunne ikke legge til bosituasjonsgrunnlaget',
     'bosituasjon.eps.alder.er.null': 'Alder til EPS er Null',
     'bosituasjon.kunne.ikke.slå.opp.eps': 'Kunne ikke slå opp EPS',
 
@@ -67,33 +54,17 @@ export default {
     'formue.periode.utenfor.bosituasjonsperiode': 'Ikke lov med formueperiode utenfor bosituasjonperioder',
     'formue.ikke.lov.eps.formue.uten.eps': 'Ikke lov med formue for EPS hvis søker ikke har EPS',
 
-    'fradrag.kunne.ikke.legge.til': 'Kunne ikke legge til fradragsgrunnlag',
-    'fradrag.ugyldig.type': 'Ugyldig fradragstype',
     'fradrag.kunne.ikke.lage': 'Kunne ikke lage fradrag',
     'fradrag.ikke.lov.eps.fradrag.uten.eps': 'Ikke lov med fradrag for EPS hvis søker ikke har EPS',
-    'fradrag.mangler.periode': 'Fradrag mangler periode',
 
-    'beregningOgSimulering.kan.ikke.velge.siste.måned.ved.nedgang':
+    'beregning.kan.ikke.velge.siste.måned.ved.nedgang':
         'Kan ikke velge siste måned av stønadsperioden ved nedgang i stønaden',
-    'beregningOgSimulering.ugyldig.beregnignsgrunnlag': 'Ugyldig beregningsgrunnlag',
-    'beregningOgSimulering.ikke.lov.eps.fradrag.uten.eps':
-        'Kan ikke ha fradrag knyttet til EPS når bruker ikke har EPS',
-    'beregningOgSimulering.simulering.feilet': 'Simulering feilet',
-    'beregningOgSimulering.oppdrag.stengt.eller.nede':
-        'Simulering feilet. Oppdrag/UR er stengt eller nede. Prøv på nytt eller prøv igjen i Oppdrags åpningstid.',
-    'beregningOgSimulering.finner.ikke.person': 'Simulering feilet. Finner ikke person i TPS.',
-    'beregningOgSimulering.finner.ikke.kjøretidsplan':
-        'Simulering feilet. Finner ikke kjøreplansperiode for fom-dato. Kjøreplan finnes som regel bare for inneværende år.',
-    'beregningOgSimulering.oppdraget.finnes.ikke':
-        'Simulering feilet. Oppdraget finnes ikke, brukerens transaksjoner er sannsynligvis fjernet fra Oppdrag.',
+    'beregning.ugyldig.beregnignsgrunnlag': 'Ugyldig beregningsgrunnlag',
+    'beregning.ikke.lov.eps.fradrag.uten.eps': 'Kan ikke ha fradrag knyttet til EPS når bruker ikke har EPS',
 
     'brev.navnoppslag.feilet': 'Kunne ikke hente navn for saksbehandler eller attestant',
     'brev.fant.ikke.gjeldende.utbetaling': 'Kunne ikke hente gjeldende utbetaling',
-    'brev.kunne.ikke.distribuere': 'Kunne ikke distribuere brev',
-    'brev.kunne.ikke.journalføre': 'Kunne ikke journalføre brev',
-    'brev.kunne.ikke.generere': 'Kunne ikke generere brev',
     'brev.kunne.ikke.lage': 'Kunne ikke lage brevutkast',
-    'brev.generering.dokument.feilet': 'Feil ved generering av dokument',
 
     'feil.ukjentFeil': 'Ukjent feil',
 };
