+++ src/components/Feilresponser/søknadsbehandlingFeilresponser/søknadsbehandlingFeilresponser-nb.ts
@@ -1,0 +1,18 @@
+export default {
+    'generell.feil.ved.henting.av.saksbehandler.eller.attestant': 'Feil ved henting av saksbehandler eller attestant',
+    'generell.finner.ikke.utbetaling': 'Finner ikke utbetaling',
+    'generell.attestant.samme.som.saksbehandler': 'Attestant er samme som saksbehandler',
+    'generell.mangler.begrunnelse': 'Mangler begrunnelse',
+
+    'stønadsperiode.oppdatering.av.periode': 'Feil ved oppdatering av stønadsperiode',
+    'stønadsperiode.periode.før.2021': 'Stønadsperiode kan ikke starte før Januar 2021',
+    'stønadsperiode.periode.maks.12.måneder': 'En stønadsperiode kan være maks 12 måneder',
+
+    'formue.ugyldige.verdier': 'Ugyldige formue verdier',
+    'formue.har.ikke.ektefelle': 'Har ikke ektefelle',
+
+    'fradrag.ugyldig.type': 'Ugyldig fradragstype',
+    'fradrag.mangler.periode': 'Fradrag mangler periode',
+
+    'feil.ukjentFeil': 'Ukjent feil',
+};
