+++ src/pages/saksbehandling/søknadsbehandling/fast-opphold-i-norge/FastOppholdINorge.tsx
@@ -1,13 +1,13 @@
 import * as RemoteData from '@devexperts/remote-data-ts';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { Eq } from 'fp-ts/lib/Eq';
-import AlertStripe from 'nav-frontend-alertstriper';
 import { Feiloppsummering, Radio, RadioGruppe, Textarea } from 'nav-frontend-skjema';
 import NavFrontendSpinner from 'nav-frontend-spinner';
 import React, { useRef } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { useHistory } from 'react-router-dom';
 
+import Feilresponser from '~components/Feilresponser/Feilresponser';
 import { FastOppholdFaktablokk } from '~components/oppsummering/vilkårsOppsummering/faktablokk/faktablokker/FastOppholdFaktablokk';
 import ToKolonner from '~components/toKolonner/ToKolonner';
 import * as sakSlice from '~features/saksoversikt/sak.slice';
@@ -148,11 +148,7 @@ const FastOppholdINorge = (props: VilkårsvurderingBaseProps) => {
                             RemoteData.fold(
                                 () => null,
                                 () => <NavFrontendSpinner>{formatMessage('display.lagre.lagrer')}</NavFrontendSpinner>,
-                                () => (
-                                    <AlertStripe type="feil">
-                                        {formatMessage('display.lagre.lagringFeilet')}
-                                    </AlertStripe>
-                                ),
+                                (err) => <Feilresponser error={err} />,
                                 () => null
                             )
                         )}
