+++ src/pages/saksbehandling/søknadsbehandling/flyktning/Flyktning.tsx
@@ -8,6 +8,7 @@ import React, { useMemo, useRef } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { useHistory } from 'react-router-dom';
 
+import Feilresponser from '~components/Feilresponser/Feilresponser';
 import { FlyktningFaktablokk } from '~components/oppsummering/vilkårsOppsummering/faktablokk/faktablokker/FlyktningFaktablokk';
 import ToKolonner from '~components/toKolonner/ToKolonner';
 import * as sakSlice from '~features/saksoversikt/sak.slice';
@@ -207,11 +208,7 @@ const Flyktning = (props: VilkårsvurderingBaseProps) => {
                             RemoteData.fold(
                                 () => null,
                                 () => <NavFrontendSpinner>{formatMessage('display.lagre.lagrer')}</NavFrontendSpinner>,
-                                () => (
-                                    <AlertStripe type="feil">
-                                        {formatMessage('display.lagre.lagringFeilet')}
-                                    </AlertStripe>
-                                ),
+                                (err) => <Feilresponser error={err} />,
                                 () => null
                             )
                         )}
